+++ whitelist-authenticator/src/main/java/de/intension/authentication/WhitelistAuthenticator.java
@@ -12,12 +12,14 @@ import org.keycloak.authentication.AuthenticationFlowContext;
 import org.keycloak.authentication.AuthenticationFlowError;
 import org.keycloak.authentication.Authenticator;
 import org.keycloak.authentication.authenticators.broker.AbstractIdpAuthenticator;
+import org.keycloak.authentication.authenticators.broker.util.PostBrokerLoginConstants;
 import org.keycloak.authentication.authenticators.broker.util.SerializedBrokeredIdentityContext;
 import org.keycloak.constants.AdapterConstants;
 import org.keycloak.models.KeycloakSession;
 import org.keycloak.models.RealmModel;
 import org.keycloak.models.UserModel;
 import org.keycloak.services.ErrorPage;
+import org.keycloak.services.resources.LoginActionsService;
 import org.keycloak.utils.StringUtil;
 
 import de.intension.authentication.rest.IdPAssignmentsClient;
@@ -29,7 +31,7 @@ public class WhitelistAuthenticator
     implements Authenticator, IdpHintParamName
 {
 
-    private static final Logger  logger = Logger.getLogger(WhitelistAuthenticator.class);
+    private static final Logger        logger = Logger.getLogger(WhitelistAuthenticator.class);
     private final IdPAssignmentsClient client;
 
     public WhitelistAuthenticator(IdPAssignmentsClient client)
@@ -59,20 +61,33 @@ public class WhitelistAuthenticator
     private String getProviderIdFromContext(AuthenticationFlowContext context)
     {
         String idpHintParamName = getIdpHintParamName(context);
-        String providerId = context.getUriInfo().getQueryParameters().getFirst(idpHintParamName);
-        providerId = StringUtil.isNotBlank(providerId) ? providerId
-                : context.getUriInfo().getQueryParameters().getFirst(AdapterConstants.KC_IDP_HINT);
-        if (StringUtil.isBlank(providerId)) {
-            try {
+
+        String flowPath = context.getFlowPath();
+        String providerId = null;
+        try {
+            if (LoginActionsService.AUTHENTICATE_PATH.equals(flowPath)) {
+                providerId = context.getUriInfo().getQueryParameters().getFirst(idpHintParamName);
+                providerId = StringUtil.isNotBlank(providerId) ? providerId
+                        : context.getUriInfo().getQueryParameters().getFirst(AdapterConstants.KC_IDP_HINT);
+            }
+            else if (LoginActionsService.FIRST_BROKER_LOGIN_PATH.equals(flowPath)) {
                 SerializedBrokeredIdentityContext serializedCtx = SerializedBrokeredIdentityContext
                     .readFromAuthenticationSession(context.getAuthenticationSession(),
                                                    AbstractIdpAuthenticator.BROKERED_CONTEXT_NOTE);
                 if (serializedCtx != null) {
                     providerId = serializedCtx.getIdentityProviderId();
                 }
-            } catch (Exception e) {
-                logger.warn(e.getLocalizedMessage());
             }
+            else if (LoginActionsService.POST_BROKER_LOGIN_PATH.equals(flowPath)) {
+                SerializedBrokeredIdentityContext serializedCtx = SerializedBrokeredIdentityContext
+                    .readFromAuthenticationSession(context.getAuthenticationSession(),
+                                                   PostBrokerLoginConstants.PBL_BROKERED_IDENTITY_CONTEXT);
+                if (serializedCtx != null) {
+                    providerId = serializedCtx.getIdentityProviderId();
+                }
+            }
+        } catch (Exception e) {
+            logger.warn(e.getLocalizedMessage());
         }
         return providerId;
     }
@@ -147,15 +162,17 @@ public class WhitelistAuthenticator
         if (config.containsKey(configKey)) {
             value = config.get(configKey);
         }
-        else if(defaultValue != null){
+        else if (defaultValue != null) {
             value = defaultValue;
-        } else {
+        }
+        else {
             logger.errorv("Provider %s - Parameter %s must not be null", WhitelistAuthenticatorFactory.PROVIDER_ID, configKey);
         }
         return value;
     }
 
-    public IdPAssignmentsClient getClient(){
+    public IdPAssignmentsClient getClient()
+    {
         return client;
     }
 }